Create separate packages for interface and compile using lerna
