+++ .travis.yml
@@ -4,6 +4,8 @@ node_js:
   - "7"
   - "6"
 before_script:
+  - './node_modules/.bin/lerna bootstrap'
+  - './node_modules/.bin/lerna run build'
   - 'if [ -n "${REACT-}" ] && [ "${TEST-}" = true ]; then sh install-relevant-react.sh && npm ls >/dev/null; fi'
 script:
   - 'if [ -n "${LINT-}" ]; then npm run lint; fi'
