--- compile.js
@@ -1,2 +1,0 @@
-// eslint-disable-next-line import/no-unresolved
-module.exports = require('./dist/compile').default;
