+++ packages/compile/.babelrc
@@ -1,0 +1,3 @@
+{
+  "presets": ["airbnb"]
+}
