+++ packages/compile/.eslintrc
@@ -1,0 +1,8 @@
+{
+  "root": true,
+  "extends": "airbnb-base",
+  "env": {
+    "browser": true,
+    "node": true
+  }
+}
