+++ packages/compile/src/cli.js
@@ -2,7 +2,7 @@ import { writeFile } from 'fs';
 import { resolve } from 'path';
 import CleanCSS from 'clean-css';
 
-import compileCSS from './compile';
+import compileCSS from './';
 
 const entryPointFilePath = process.argv[2];
 const CSS = compileCSS(entryPointFilePath);
