+++ packages/compile/test/ComponentA.jsx
