+++ packages/compile/test/ComponentB.jsx
