+++ packages/compile/test/entryPoint.jsx
