+++ packages/compile/test/withStyles.js
@@ -1,6 +1,6 @@
 import { withStyles } from 'react-with-styles';
 import ThemedStyleSheet from 'react-with-styles/lib/ThemedStyleSheet';
-import CSSInterface from '../src';
+import CSSInterface from 'react-with-styles-interface-css';
 
 ThemedStyleSheet.registerTheme({});
 ThemedStyleSheet.registerInterface(CSSInterface);
