+++ packages/interface/index.js
@@ -1,0 +1,2 @@
+// eslint-disable-next-line import/no-unresolved
+module.exports = require('./dist/index.js').default;
