+++ packages/interface/jest.config.js
@@ -1,0 +1,4 @@
+module.exports = {
+  testEnvironment: 'node',
+  testRegex: '.*(\\.|/|_)(test)\\.jsx?$',
+};
