+++ packages/interface/src/utils/registerMaxSpecificity.js
