+++ packages/interface/test/.eslintrc
@@ -1,0 +1,10 @@
+{
+  "env": {
+    "jest": true
+  },
+  "rules": {
+    "import/no-extraneous-dependencies": [2, {
+      "devDependencies": true
+    }],
+  }
+}
